Only recreate symlink if there isn't already a similarly named file/directory in new directory
git-svn-id: https://svn.calendarserver.org/repository/calendarserver/CalendarServer/trunk@4179 e27351fd-9f3e-4f54-a53b-843176b1656c

+++ contrib/migration/59_calendarmigrator.py
@@ -133,7 +133,7 @@ def migrateConfiguration(options):
             oldPath = os.path.join(oldConfigDir, name)
             newPath = os.path.join(newConfigDir, name)
 
-            if os.path.islink(oldPath):
+            if os.path.islink(oldPath) and not os.path.exists(newPath):
                 # Recreate the symlink if it won't overwrite an existing file
                 link = os.readlink(oldPath)
                 os.symlink(link, newPath)
